requirements.txt
pymysql
requests
redis
pydub
django
djangorestframework
uwsgi